Latest Improved Code Quality Backend Code

+++ src/main/java/com/prodapt/capstoneproject/model/AdminActionsReport.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class AdminActionsReport {
 	private Long adminid;
-	private Integer adminactions;
-	private Integer reactivations;
+	private Long adminactions;
+	private Long reactivations;
 
 }
